Create Service layer Unit test

+++ Task 2/Data/DataAPI/IProduct.cs
@@ -3,7 +3,7 @@
     public interface IProduct
     {
          int ProductID { get; set; }
-         double Price { get; set; }
+         decimal Price { get; set; }
          string Category { get; set; }
     }
